+++ test/apps/test_deeplabv3p_ar_detect_app.sh
@@ -24,13 +24,22 @@ set -x
 ${launcher} ${app_prefix}/teca_deeplabv3p_ar_detect              \
     --input_regex "${data_root}/ARTMIP_MERRA_2D_2017-05.*\.nc$"  \
     --pytorch_deeplab_model ${data_root}/cascade_deeplab_IVT.pt  \
-    --n_threads ${n_threads}                                     \
-    --output_file test_deeplabv3p_ar_detect_app_output_%t%.nc
+    --output_file test_deeplabv3p_ar_detect_app_output_%t%.nc    \
+    --steps_per_file 365 --n_threads ${n_threads}
 
-# run the diff
-${app_prefix}/teca_cartesian_mesh_diff                           \
-    "${data_root}/test_deeplabv3p_ar_detect_app_ref.nc"          \
-    "test_deeplabv3p_ar_detect_app_output.*\.nc"
+do_test=1
+if [[ ${do_test} -eq 0 ]]
+then
+    # regenerate the baseline
+    cp test_deeplabv3p_ar_detect_app_output*.nc \
+        ${data_root}/test_deeplabv3p_ar_detect_app_ref.nc
+else
+    # run the diff
+    ${app_prefix}/teca_cartesian_mesh_diff                                      \
+        --reference_dataset "${data_root}/test_deeplabv3p_ar_detect_app_ref.nc" \
+        --test_dataset "test_deeplabv3p_ar_detect_app_output.*\.nc"             \
+        --arrays ar_probability ar_binary_tag --test_tolerance 1e-3 --verbose
 
-# clean up
-rm test_deeplabv3p_ar_detect_app_output*.nc
+    # clean up
+    rm test_deeplabv3p_ar_detect_app_output*.nc
+fi
